refactor: Add a reducer-types directoy to store reducer code
#116
