+++ packages/data-point/lib/entity-types/entity-hash/factory.js
@@ -1,8 +1,8 @@
 'use strict'
 
 const _ = require('lodash')
-const createReducer = require('../../reducer').create
-const createReducerObject = require('../../reducer-object').create
+const createReducer = require('../../reducer-types').create
+const createReducerObject = require('../../reducer-types/reducer-object').create
 const deepFreeze = require('deep-freeze')
 const parseCompose = require('../parse-compose')
 const createBaseEntity = require('../base-entity').create
