+++ packages/data-point/lib/reducer-types/reducer-entity/factory.test.js
