+++ packages/data-point/lib/reducer-types/reducer-function/factory.test.js
