+++ packages/data-point/lib/reducer-types/reducer-function/index.js
