+++ packages/data-point/lib/reducer-types/reducer-list/factory.js
