+++ packages/data-point/lib/reducer-types/reducer-list/resolve.js
