+++ packages/data-point/lib/reducer-types/reducer-object/factory.js
