+++ packages/data-point/lib/reducer-types/reducer-object/index.js
