+++ packages/data-point/lib/reducer-types/reducer-path/factory.js
