+++ packages/data-point/lib/reducer-types/reducer-path/index.js
